+++ modmail/config.py
@@ -7,8 +7,10 @@ from typing import Any, Dict, Tuple
 import discord
 import toml
 from pydantic import BaseSettings as PydanticBaseSettings
+from pydantic.color import Color as ColorBase
 from pydantic.env_settings import SettingsSourceCallable
 from pydantic.types import conint
+
 
 log = logging.getLogger(__name__)
 
@@ -107,6 +109,16 @@ class BotMode(BaseSettings):
     develop: bool = False
 
 
+class Colors(BaseSettings):
+    """
+    Default colors.
+
+    These should only be changed here to change the default colors.
+    """
+
+    embed_color: ColorBase = "0087BD"
+
+
 class DevConfig(BaseSettings):
     """
     Developer specific configuration.
@@ -124,6 +136,7 @@ class ThreadConfig(BaseSettings):
 
 class ModmailConfig(BaseSettings):
     bot: BotConfig
+    colors: Colors
     dev: DevConfig
     thread: ThreadConfig