+++ docs/changelog.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - At a later point in time, this will be additionally included in a configuration command.
 
 ### Added
+
+- Interaction Paginator that uses discord buttons (#50)
 - docker-compose.yml (#13)
     - Running the bot after configuring the env vars is now as simple as `docker-compose up`
 - Automatic docker image creation: `ghcr.io/discord-modmail/modmail` (#19)
@@ -28,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `bot.start()` can also be used if already in a running event loop. Keep in mind using it will require
         handling loop errors, as run() does this automatically.
 
+### Internal
+
+- Code style: two blank lines after imports instead of one. (#70)
 
 
 ## [0.1.0] - 2021-08-13